Blog_Website/requirement.txt
amqp==1.4.9
anyjson==0.3.3
backcall==0.1.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.7.1
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
decorator==4.0.10
Django==1.11.20
django-crispy-forms==1.7.2
django-paypal==0.2.5
djangorestframework==3.9.2
djangorestframework-stubs==0.4.2
docopt==0.6.2
enum34==1.1.6
gunicorn==19.9.0
idna==2.8
ipython==7.5.0
ipython-genutils==0.1.0
jedi==0.13.3
kombu==3.0.37
mysqlclient==1.4.2.post1
p==0.0.0
parso==0.4.0
pathlib2==2.1.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==5.4.1
Pillow-PIL==0.1.dev0
prompt-toolkit==2.0.9
psycopg2==2.7.7
ptpython==2.0.4
ptyprocess==0.5.1
PyAudio==0.2.11
pycryptodome==3.8.1
Pygments==2.4.0
pyttsx3==2.7
pytz==2018.9
pywin32==224
requests==2.21.0
six==1.12.0
soupsieve==1.9.1
South==1.0.2
SpeechRecognition==3.8.1
traitlets==4.3.2
urllib3==1.24.2
vine==1.3.0
wcwidth==0.1.7
whitenoise==4.1.2
wikipedia==1.4.0
